moved eddsa signing into its own file
